Test for Java assertion
When creating grammar from a set of XSDs provided by the XMPP Standards Foundation,
`com.siemens.ct.exi.grammars.EXIContentModelBuilder.handleStateEntries()` throws an
AssertionError (when assertions are enabled in the VM), with the latest state of
the code (v1.0.4 and 1.0.5-SNAPSHOT).

This commit adds a unit test that reproduces this behavior.